Add undo button functionality
The undo button allows users to go back on a digit entered. This is
different from the clear button because it only deletes one character at
a time.

The undo button icon now fits the container better. Before, the undo
arrow graphic was an SVG element in HTML, but it would cause clicks on
the SVG element to be ignored. Now, the undo button graphic is a
background image with a size set relative to the font size.